Update 1.1.2
Improved ServiceId comparison logic. Now Type is allowed to be null, which allows to get services that satisfy the key and have any type.

ServiceId can now be fully deconstructed into its components (Type and Key). This allows for more flexible scenarios and is a more readable variation of the implicit operators previously used to cast ServiceId to Type

Minor optimizations and refactoring.